requirements.txt
# RAG pipeline
chromadb>=0.5.0,<0.6.0
sentence-transformers>=3.0.0,<4.0.0
tqdm>=4.66.0

# LLM clients (install as needed)
anthropic>=0.40.0
openai>=1.50.0

# Data processing
pandas>=2.0.0
numpy>=1.26.0

# Experiment
scikit-learn>=1.4.0
scipy>=1.12.0

# HTTP client
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
